backend/app/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
google-generativeai
pypdfium2
markdownify
Pillow
pytesseract
pdf2image
python-multipart